New architecture and go-to definition
Let the user configure a "load order", mimicking the files and directories they pass to their source port. This gets used to build ZScript include trees and so on.

To prove that this code is at least robust, it is now possible to go-to the definitions of user-declared classes, mixin classes, structs, enums, and symbolic constants (as long as they're declare at the very top level).

Next steps are to:
- Optimize symbol table generation and storage. The current solution works but isn't ideal.
- Decide how to deal with ZDoom's case-insensitive symbol names.
- Evolve the server's semantic knowledge about the source further, starting with being able to deal with types nested within types.