Start migrating away from Salsa
I'm concluding that this is not the right solution for my task. Using a monolithic global symbol table is non-negotiable due to the weaknesses in the compilation models of ZScript and DECORATE, but this runs afoul of Salsa's rules about not letting external mutable state affect the results of incremental computations. The only other option would be rebuilding the whole table in a derived query, but this would be an unacceptable performance cost. DoomLS will simply use something custom, and attempt incrementality in its own way.

This is going to be a long and tedious process, so start small; break out src/zpath.rs and link against `slotmap`.