Proof-of-concept for new architecture
Based on https://matklad.github.io/2023/05/06/zig-language-server-and-cancellation.html.

This commit is *much* bigger than I would like but I had to get pretty far to verify that it didn't follow a direction that would prove unsuitable in later commits.

What's next is to implement the rest of the ZScript frontend to determine how fast it can be, as this will give a clue as to how much time a semantic update will take after all other relevant languages are supported.